Keep Trash v2.0
With v2.0, we are moving towards a completely re-written implementation
of the Xposed module. A more robust and flexible implementation.

Google Keep has moved towards a completely material UI. The previous implementation
of the module no longer works because the code has the menu items hard coded.
Hence, replacing the action bar menu xml resources doesn't work as in the code
they are referencing the menu items to change their properties, resulting in a crash.

- We are now changing the properties of the menu items by hooking through the code,
instead of replacing the menu resource xml
- Add support for Google Keep v3.3+
- New customization available for 3 additional menu items (Color Changer, Label and Reminder)
- New app icon
- Updated Copyright
- New material drawables for the Google Keep action bar
- New material design for Keep Trash app
- Removed floating action button (for now?)
- Add Donate preference